Backend/ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
celery[redis]>=5.3.6
pillow>=10.0.0
ultralytics==8.3.0
torch>=2.0.0
torchvision>=0.15.0
